Betty style

+++ 0x00-hello_world/6-size.c
@@ -16,8 +16,7 @@ int main(void)
 	printf("Size of a char: %lu byte(s)\n", (unsigned char)sizeof(c));
 	printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(i));
 	printf("Size of a long int: %lu byte(s)\n", (unsigned long int)sizeof(d));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long long int)sizeof(e));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned int)sizeof(e));
 	printf("Size of a float: %lu byte(s)\n", (unsigned float)sizeof(f));
-	
 	return (0);
 }
